+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/viewmodels/DebugViewModel.kt
@@ -153,4 +153,10 @@ class DebugViewModel @VMInject constructor(
             fapInstallationAllApi.installAll()
         }
     }
+
+    fun brokeBytes() {
+        viewModelScope.launch {
+            serviceProvider.getServiceApi().requestApi.sendTrashBytesAndBrokeSession()
+        }
+    }
 }